feat: implemented PUT endpoint of contas_a_pagar_e_receber with unit test

+++ FastApiZeroAteDeploy/contas_a_pagar_e_receber/models/contas_a_pagar_e_receber.py
@@ -1,9 +1,10 @@
 from sqlmodel import Field, SQLModel
+from decimal import Decimal
 from typing import Optional
 
 class ContasAPagarReceber(SQLModel, table=True):
     __tablename__ = 'contas_a_pagar_e_receber' # type: ignore
     id: int | None = Field(default=None, primary_key=True)
     descricao: str
-    valor: float
+    valor: Decimal
     tipo: str
